refactoring controller
