+++ src/main/java/ifpb/controllers/interfaces/IController.java
@@ -1,14 +1,14 @@
-package ifpb.controller.interfaces;
+package ifpb.controllers.interfaces;
 
 import ifpb.Article;
 import ifpb.entitybasic.interfaces.IID;
 import ifpb.entitycomplex.interfaces.INews;
 
-import java.util.ArrayList;
+import java.util.List;
 
 public interface IController {
     INews getNewsById(IID<Integer> id);
     Article getArticleById(IID<Integer> id);
-    ArrayList<INews> getNews();
-    ArrayList<Article> getArticles();
+    List<INews> getNews();
+    List<Article> getArticles();
 }
